Fix and use Region serializer
